+++ frontend/app/(core)/billing/_lib/billing-copy.ts
@@ -1,0 +1,143 @@
+export const DEFAULT_BILLING_COPY = {
+  title: 'Billing',
+  hero: {
+    title: 'Add credits',
+    subtitle: 'Choose an amount, then complete payment in secure Stripe Checkout.',
+    liveMode: 'Live Mode',
+    testMode: 'Test Mode',
+  },
+  wallet: {
+    title: 'Add credits',
+    description: 'Your balance pays for successful renders only. Unused credits stay in your wallet.',
+    balanceLabel: 'Current balance',
+    addFunds: '{amount}',
+    quickAmount: '{amount}',
+    selectedAmount: 'Selected amount',
+    checkoutCta: 'Continue to secure Stripe Checkout · {amount}',
+    checkoutNote: 'Taxes and receipt are finalized by Stripe before payment.',
+    customLabel: 'Custom amount',
+    customPresetLabel: 'Custom',
+    customPresetHint: 'Enter amount',
+    customPlaceholder: 'Minimum $10',
+    customHint: 'Enter any amount from $10.',
+    customCta: 'Set amount',
+    customSelectCta: 'Set amount',
+    customInvalid: 'Enter a valid amount.',
+    customMin: 'Minimum is $10.',
+    chooseAmount: 'Select',
+    expressTitle: 'Fast pay options',
+    expressSubtitle: 'Apple Pay, Google Pay, PayPal, and Link appear when available on this device.',
+    expressLoading: 'Checking fast pay options…',
+    expressUnavailable: 'Fast pay is not available in this browser.',
+    expressError: 'Express checkout unavailable.',
+    expressClosed: 'Fast pay window closed. No charge made.',
+    expressAriaLabel: 'Express checkout',
+    expressRevealCta: 'Show Apple Pay / Google Pay / Link',
+    expressRevealAction: 'Show',
+    captchaPrompt: 'Security check required before Checkout.',
+    captchaComplete: 'Security check complete. Continue to payment.',
+    captchaError: 'Security check unavailable. Try again.',
+    rateLimited: 'Too many payment attempts. Try again in {seconds}s.',
+    autoTopUp: 'Enable auto top-up (optional)',
+    lowBalance: 'Your balance is low. Top up to keep creating.',
+    currencyLabel: 'Charge currency',
+    currencyLoading: 'Detecting currencies…',
+    currencyLoadError: 'Unable to load currencies. Defaulting to USD.',
+    currencyDetected: 'Auto-detected {currency}',
+    currencyOverride: 'Charging in {selected} (auto: {detected})',
+    quoteLoading: 'Fetching live totals…',
+    quoteError: 'Unable to fetch FX quotes. Amount finalized at checkout.',
+  },
+  membership: {
+    title: 'Member Status',
+    description:
+      'Status is automatic — no subscription. Calculated on a rolling 30 days of spend. Status updates daily based on your last 30 days of spend.',
+    defaultTier: 'Member',
+    savingsChip: 'You save {percent}%',
+    defaultLine: '{tier} — standard rate on every render.',
+    thresholdLine: '{tier} — spend {amount} / 30 days to save {percent}% on every eligible render.',
+    labels: {
+      member: 'Member',
+      plus: 'Plus',
+      pro: 'Pro',
+    },
+  },
+  teams: {
+    title: 'For Teams',
+    description: 'Shared wallet & roles. Let your team create with one balance. Set soft/hard project budgets. Daily summary by email.',
+    actions: {
+      invite: 'Invite teammates',
+      budgets: 'Set project budgets',
+    },
+    note: 'Team note: On shared wallets, status applies to all members.',
+    comingSoon:
+      'Coming soon — unified wallets, role-based approvals, and budgeting controls. Join the beta at {email}.',
+    contactEmail: 'support@maxvideoai.com',
+    statusLive: 'Live',
+    statusSoon: 'Coming soon',
+  },
+  receipts: {
+    title: 'Payment history',
+    subtitle: 'Wallet top-ups, render charges, refunds, and Stripe receipts where available.',
+    empty: 'No receipts yet.',
+    loading: 'Loading…',
+    loadMore: 'Load more',
+    exportCsv: 'Export CSV',
+    collapsedLabel: 'Show',
+    expandedLabel: 'Hide',
+    typeLabels: {
+      charge: 'Charge',
+      refund: 'Refund',
+      topup: 'Top-up',
+    },
+    fields: {
+      total: 'Total',
+      tax: 'Tax',
+      discount: 'Discount',
+      document: 'Document',
+    },
+    viewReceipt: 'View receipt',
+    viewInvoice: 'View invoice',
+    receiptLabel: 'Receipt',
+    invoiceLabel: 'Invoice',
+    contactSupport: 'Contact support',
+  },
+  refunds: {
+    title: 'Protections',
+    points: [
+      'Failed renders are refunded automatically.',
+      'Receipts include itemized tax and payment details.',
+      'No subscription. Add credits only when you need them.',
+    ],
+  },
+  faq: {
+    title: 'Micro-FAQ',
+    entries: [
+      { question: 'Do I need a subscription?', answer: 'No. Pay as you go.' },
+      { question: 'Can I buy just one small render?', answer: 'Yes — funds are debited per run from your $10 Starter balance.' },
+      { question: 'Will my credits expire?', answer: 'Credits don’t expire while your account remains active.' },
+      { question: 'How do discounts work?', answer: 'Member status applies automatically based on your last 30 days of spend.' },
+    ],
+    footnote:
+      'VAT included where applicable. Refunds on failed renders. “Works with” indicates compatibility; trademarks belong to their respective owners.',
+  },
+  toasts: {
+    success: 'Payment successful. Funds added to your wallet.',
+    cancelled: 'Checkout closed. No charge made.',
+  },
+  errors: {
+    loadReceipts: 'We couldn’t load your payment history. Refresh the page or contact support.',
+    loadMore: 'Failed to load more receipts',
+    lowBalance: 'Your balance is low. Top up to keep creating.',
+    topupStart: 'Unable to start payment.',
+  },
+  authGate: {
+    title: 'Create an account to fund your wallet',
+    body: 'You can browse Billing and inspect live pricing, but adding funds requires an account.',
+    primary: 'Create account',
+    secondary: 'Sign in',
+    close: 'Maybe later',
+  },
+};
+
+export type BillingCopy = Omit<typeof DEFAULT_BILLING_COPY, 'estimator'>;
